backend_sentify/requirements.txt
Flask
flask-cors
google-api-python-client
nltk
python-dotenv